+++ source/ui/ProgramManagementTab.cpp
@@ -63,8 +63,8 @@ ProgramManagementTab::ProgramManagementTab(MidiHandler *midiHandler) : juce::Com
             juce::MemoryBlock memBlock;
             importFile.loadFileAsData(memBlock);
 
-            size_t start = 0;
-            for ( size_t i = 0; i < (size_t)memBlock.getSize(); i++ )
+            int start = 0;
+            for ( int i = 0; i < (int)memBlock.getSize(); i++ )
             {
                 if ( i == start && (uint8_t)memBlock[i] != 0xF0 )
                 {
